db/requirements.txt
psycopg2-binary
shapely
